factor out htmltagfixer and fix lint
