+++ overcast_to_sqlite/html/page.py
@@ -1,10 +1,10 @@
 import html
 import re
 from pathlib import Path
-from html.parser import HTMLParser
 
 from overcast_to_sqlite.constants import DESCRIPTION
 from overcast_to_sqlite.datastore import Datastore
+from overcast_to_sqlite.html.htmltagfixer import HTMLTagFixer
 
 
 def _convert_urls_to_links(text: str) -> str:
@@ -27,54 +27,16 @@ def _convert_urls_to_links(text: str) -> str:
     return "".join(result)
 
 
-class HTMLTagFixer(HTMLParser):
-    SELF_CLOSING_TAGS = {'br', 'hr', 'img', 'input', 'meta', 'link', 'area', 'base', 'col', 'embed', 'source', 'track', 'wbr'}
-
-    def __init__(self):
-        super().__init__()
-        self.open_tags = []
-        self.output = []
-
-    def handle_starttag(self, tag, attrs):
-        # Self-closing tags don't need closing tags
-        attr_str = ''.join(f' {name}="{value}"' for name, value in attrs)
-        self.output.append(f'<{tag}{attr_str}>')
-
-        if tag.lower() not in self.SELF_CLOSING_TAGS:
-            self.open_tags.append(tag)
-    
-    def handle_endtag(self, tag):
-        self.output.append(f'</{tag}>')
-        if self.open_tags and self.open_tags[-1] == tag:
-            self.open_tags.pop()
-    
-    def handle_data(self, data):
-        self.output.append(data)
-    
-    def handle_entityref(self, name):
-        self.output.append(f'&{name};')
-    
-    def handle_charref(self, name):
-        self.output.append(f'&#{name};')
-    
-    def get_fixed_html(self):
-        # Add closing tags for any unclosed tags in reverse order
-        for tag in reversed(self.open_tags):
-            self.output.append(f'</{tag}>')
-        return ''.join(self.output)
-
-
 def _fix_unclosed_html_tags(html_string: str) -> str:
     """Fix unclosed HTML tags by adding missing closing tags."""
     if not html_string.strip():
         return html_string
-        
+
     fixer = HTMLTagFixer()
     try:
         fixer.feed(html_string)
         return fixer.get_fixed_html()
-    except Exception:
-        # If parsing fails, return original string
+    except Exception:  # noqa: BLE001
         return html_string
 
 
@@ -93,7 +55,9 @@ def generate_html_played(db_path: str, html_output_path: Path) -> None:
     last_user_updated_date = None
     for ep in episodes:
         ep["episode_title"] = html.escape(ep["episode_title"])
-        ep[DESCRIPTION] = _fix_unclosed_html_tags(_convert_urls_to_links(ep[DESCRIPTION]))
+        ep[DESCRIPTION] = _fix_unclosed_html_tags(
+            _convert_urls_to_links(ep[DESCRIPTION]),
+        )
         user_date = ep["userUpdatedDate"].split("T")[0]
         if last_user_updated_date != user_date:
             page_vars["episodes"] += (
