+++ pyproject.toml
@@ -45,6 +45,15 @@ overcast-to-sqlite = "overcast_to_sqlite.cli:cli"
 [project.urls]
 Homepage = "https://github.com/hbmartin/overcast-to-sqlite"
 
+[dependency-groups]
+dev = [
+    "black>=25.1.0",
+    "mypy>=1.16.0",
+    "ruff>=0.11.13",
+    "types-python-dateutil>=2.9.0.20250516",
+    "types-requests>=2.32.4.20250611",
+]
+
 [tool.ruff]
 # Exclude a variety of commonly ignored directories.
 exclude = [".bzr", ".direnv", ".eggs", ".git", ".git-rewrite", ".hg", ".ipynb_checkpoints", ".mypy_cache", ".nox", ".pants.d", ".pyenv", ".pytest_cache", ".pytype", ".ruff_cache", ".svn", ".tox", ".venv", ".vscode", "__pypackages__", "_build", "buck-out", "build", "dist", "node_modules", "setup.py", "site-packages", "venv"]
@@ -55,12 +64,13 @@ indent-width = 4
 target-version = "py311"
 
 lint.select = ["ALL"]
-lint.ignore = ["D100", "D101", "D103", "D104", "D105", "D107", "D203", "D213", "ICN001", "S608", "S113", "S311", "S314", "T201"]
+lint.ignore = ["D100", "D101", "D102", "D103", "D104", "D105", "D107", "D203", "D213", "ICN001", "S608", "S113", "S311", "S314", "T201"]
 
 [tool.ruff.lint.per-file-ignores]
 "overcast_to_sqlite/cli.py" = ["FBT001"]
 "overcast_to_sqlite/more_itertools.py" = ["RET505", "EM101", "TRY003", "ANN202", "D415", "D400", "ANN001", "FBT002", "ANN201"]
 "overcast_to_sqlite/episode.py" = ["SIM102"]
+"overcast_to_sqlite/html/htmltagfixer.py" = ["ANN001", "ANN201", "ANN204"]
 
 [tool.ruff.format]
 # Same as Black.
